Set the 'title' attribute of the timeago element to be the text from the element.
(Only happens the first time we wipe the original text, and only if a
'title' wasn't already set).

This allows your original HTML (before JS runs) to have a human readable date like
"April 11, 2014" and then have smart-time-ago make it say "1 month ago"
but then the user can still hover their mouse to see the more detailed/accurate
format of the date.
<time class="timeago" datetime="2014-04-11">April 11, 2014</time>

Of course, a tooltip could be done already without these changes if you make the
original HTML like this:
<time class="timeago" datetime="2014-04-11" title="April 11, 2014">April 11, 2014</time>
However having the 'title' before the text has changed is redundant, and not in the
spirit of JS gracefully enhancing the page.